requirement.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
PyJWT
pwdlib[argon2]
jinja2
python-multipart
aiofiles
python-dotenv
psycopg2-binary

# --- HOW TO RUN THIS PROJECT ---
# 1. Activate environment:
#    In PowerShell: .\env\Scripts\Activate.ps1
#    In Git Bash:   source env/Scripts/activate
#
# 2. Run Migration (Required once for new features):
#    python migrate_add_columns.py
#
# 3. Start Server:
#    python main.py
#    (OR: uvicorn main:app --reload)